pydrake: Rename {drawing => visualization.plotting}
