+++ bindings/pydrake/visualization/all.py
@@ -1,0 +1,1 @@
+from .plotting import *
