+++ bindings/pydrake/visualization/plotting.py
